+++ src/app.d.ts
@@ -7,22 +7,23 @@ declare namespace App {
   }
 
   interface PageData {
+    csrfHeaderKey: string
+    csrfParamKey: string
+    csrfToken: string
     fetch: Fetch
-    session: Session
+  }
+
+  interface PageError {
+    message: string
   }
 
   interface Platform {}
 
   interface Session {
-    csrfHeaderKey?: string
-    csrfParamKey?: string
-    csrfToken?: string
+    csrfHeaderKey: string
+    csrfParamKey: string
+    csrfToken: string
   }
-}
-
-interface ImportMetaEnv {
-  VITE_ALLOWED_REQUEST_METHODS: string
-  VITE_SESSION_KEY: string
 }
 
 type Fetch = (info: RequestInfo, init?: RequestInit) => Promise<Response>