+++ src/hooks.client.ts
@@ -1,0 +1,5 @@
+import type { HandleClientError } from '@sveltejs/kit'
+
+export const handleError: HandleClientError = ({ error }) => {
+  if (error instanceof Error) console.log(error.message)
+}