+++ src/lib/envs/client.ts
@@ -1,6 +1,10 @@
-export const allowedRequestMethods =
-  import.meta.env.VITE_ALLOWED_REQUEST_METHODS
-    .split(',')
-    .map(method => method.trim())
+import {
+  PUBLIC_ALLOWED_REQUEST_METHODS,
+  PUBLIC_SESSION_KEY
+} from '$env/static/public'
 
-export const sessionKey = import.meta.env.VITE_SESSION_KEY
+export const allowedRequestMethods = PUBLIC_ALLOWED_REQUEST_METHODS
+  .split(',')
+  .map(method => method.trim())
+
+export const sessionKey = PUBLIC_SESSION_KEY