+++ src/routes/handlers/filter-request-methods.json/+server.ts
@@ -1,0 +1,5 @@
+import { type RequestHandler, json } from '@sveltejs/kit'
+
+export const DELETE: RequestHandler = async () => {
+  return json({ success: true })
+}