+++ src/routes/handlers/filter-request-methods/+page.server.ts
@@ -1,5 +1,0 @@
-import type { Action } from '@sveltejs/kit'
-
-export const DELETE: Action = async () => {
-  return { status: 200, errors: {} }
-}