+++ src/routes/handlers/filter-request-methods/+page.svelte
@@ -1,13 +1,18 @@
 <script lang="ts">
-  export let data: App.PageData
+  export let data: App.PageData // eslint-disable-line no-undef
 
-  const endpoint = '/handlers/filter-request-methods'
+  let { fetch } = data
+  $: ({ fetch } = data)
+
+  const endpoint = '/handlers/filter-request-methods.json'
+
   let response: Response | undefined
 
   const sendDeleteRequest = async () => {
-    response = await data.fetch(endpoint, {
+    response = await fetch(endpoint, {
       method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' }
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ data: null })
     })
   }
 </script>