+++ src/routes/handlers/verify-csrf-token.json/+server.ts
@@ -1,0 +1,5 @@
+import { type RequestHandler, json } from '@sveltejs/kit'
+
+export const POST: RequestHandler = async () => {
+  return json({ success: true })
+}