+++ src/routes/handlers/verify-csrf-token/+page.server.ts
@@ -1,5 +1,0 @@
-import type { Action } from '@sveltejs/kit'
-
-export const POST: Action = async () => {
-  return { status: 200, errors: {} }
-}