+++ src/routes/handlers/verify-csrf-token/+page.svelte
@@ -1,19 +1,19 @@
 <script lang="ts">
-  import { page } from '$app/stores'
+  export let data: App.PageData // eslint-disable-line no-undef
 
-  export let data: App.PageData
+  let { csrfHeaderKey, csrfParamKey, csrfToken, fetch } = data
+  $: ({ csrfHeaderKey, csrfParamKey, csrfToken, fetch } = data)
 
-  const { csrfHeaderKey, csrfParamKey, csrfToken } = $page.data.session
   const endpoint = '/handlers/verify-csrf-token'
+
   let response: Response | undefined
 
   const jsonWithToken = async () => {
     const headers = new Headers
     headers.set('Content-Type', 'application/json')
+    headers.set(csrfHeaderKey, csrfToken)
 
-    if (csrfHeaderKey && csrfToken) headers.set(csrfHeaderKey, csrfToken)
-
-    response = await data.fetch(endpoint, {
+    response = await fetch(`${endpoint}.json`, {
       method: 'POST',
       headers,
       body: JSON.stringify({ city: 'Kumasi' })
@@ -21,7 +21,7 @@
   }
 
   const jsonWithoutToken = async () => {
-    response = await data.fetch(endpoint, {
+    response = await fetch(`${endpoint}.json`, {
       method: 'POST',
       headers: { 'Content-Type': 'application/json' },
       body: JSON.stringify({ city: 'Kumasi' })
@@ -31,13 +31,13 @@
 
 <h1>{response?.status || 0}</h1>
 
-<form id="with-token" method="POST" action={endpoint}>
+<form id="with-token" method="POST" action="{endpoint}?/create">
   <input type="hidden" name={csrfParamKey} value={csrfToken} />
   <input type="hidden" name="city" value="Kumasi" />
   <button type="submit">Send</button>
 </form>
 
-<form id="without-token" method="POST" action={endpoint}>
+<form id="without-token" method="POST" action="{endpoint}?/create">
   <input type="hidden" name="city" value="Kumasi" />
   <button type="submit">Send</button>
 </form>